refactor(splitter): 优化目录结构
