+++ packages/devui-vue/devui/splitter/src/components/splitter-bar.scss
