+++ packages/devui-vue/devui/splitter/src/components/splitter-pane.scss
