+++ packages/devui-vue/devui/splitter/src/d-resize-directive.ts
